refactor: move MongoDBAasServiceOperations to aasservice-backend-mongodb
