+++ basyx.aasrepository/basyx.aasrepository-backend-mongodb/pom.xml
@@ -32,6 +32,10 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.digitaltwin.basyx</groupId>
+			<artifactId>basyx.aasservice-backend-mongodb</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.digitaltwin.basyx</groupId>
 			<artifactId>basyx.aasrepository-core</artifactId>
 			<classifier>tests</classifier>
 			<scope>test</scope>
