+++ basyx.aasservice/pom.xml
@@ -16,6 +16,7 @@
 	<modules>
 		<module>basyx.aasservice-core</module>
 		<module>basyx.aasservice-backend-inmemory</module>
+		<module>basyx.aasservice-backend-mongodb</module>
 		<module>basyx.aasservice-feature-mqtt</module>
 		<module>basyx.aasservice-client</module>
 	</modules>
